refactor(open-ai): create client instance in nitro plugin
