+++ src/server/types/h3.ts
@@ -1,5 +1,8 @@
+import type OpenAI from 'openai'
+
 declare module 'h3' {
   interface H3EventContext {
+    $openAi?: OpenAI
     $timezone?: string
   }
 }
