+++ src/server/utils/dependencies/index.ts
@@ -1,1 +1,2 @@
 export * from './firebase'
+export * from './openAi'
