+++ src/server/utils/dependencies/openAi.ts
@@ -1,0 +1,7 @@
+export const useOpenAi = () => {
+  const event = useEvent()
+
+  if (!event.context.$openAi) throw createError('openai')
+
+  return event.context.$openAi
+}
